utils/mobile_utils.py
from appium import webdriver

class MobileUtils:
    def __init__(self, desired_caps):
        self.driver = webdriver.Remote('http://localhost:4723/wd/hub', desired_caps)

    def find_element(self, locator):
        return self.driver.find_element_by_xpath(locator)

    def click(self, locator):
        self.find_element(locator).click()

    def send_keys(self, locator, text):
        self.find_element(locator).send_keys(text)

    def get_text(self, locator):
        return self.find_element(locator).text