requirements.txt
allure-pytest>=2.13.5
allure-python-commons>=2.13.5
annotated-types>=0.7.0
anthropic>=0.49.0
anyio>=4.8.0
Appium-Python-Client>=4.5.1
argcomplete>=3.6.0
assertpy>=1.1
attrs>=25.3.0
boto3>=1.37.13
botocore>=1.37.13
browserstack-local>=1.2.10
browserstack_sdk>=1.24.6
cachetools>=5.5.2
certifi>=2025.1.31
charset-normalizer>=3.4.1
cohere>=5.14.0
colorama>=0.4.6
Deprecated>=1.2.18
distro>=1.9.0
dukpy>=0.5.0
eval_type_backport>=0.2.2
exceptiongroup>=1.2.2
execnet>=2.1.1
Faker>=37.0.0
fastavro>=1.10.0
filelock>=3.18.0
fsspec>=2025.3.0
gherkin-official>=29.0.0
gitdb>=4.0.12
GitPython>=3.1.44
google-auth>=2.38.0
greenlet>=3.1.1
griffe>=1.6.0
groq>=0.19.0
grpcio>=1.62.3
h11>=0.14.0
httpcore>=1.0.7
httpx>=0.28.1
httpx-sse>=0.4.0
huggingface-hub>=0.29.3
idna>=3.10
importlib_metadata>=8.6.1
iniconfig>=2.0.0
Jinja2>=3.1.6
jiter>=0.9.0
jmespath>=1.0.1
jsonpath-python>=1.0.6
logfire-api>=3.8.1
Mako>=1.3.9
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
mdurl>=0.1.2
mistralai>=1.5.1
mypy-extensions>=1.0.0
openai>=1.66.3
opentelemetry-api>=1.31.0
outcome>=1.3.0.post0
packaging>=24.2
parse>=1.20.2
parse_type>=0.6.4
percy-appium-app>=2.0.6
playwright>=1.50.0
pluggy>=1.5.0
prompt_toolkit>=3.0.50
protobuf>=4.21.6
psutil>=7.0.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.1
pydantic>=2.10.6
pydantic-ai>=0.0.40
pydantic-ai-slim>=0.0.40
pydantic-graph>=0.0.40
pydantic_core>=2.27.2
pyee>=12.1.1
Pygments>=2.19.1
pyobjc-core>=11.0
pyobjc-framework-Cocoa>=11.0
pyobjc-framework-SystemConfiguration>=11.0
PyPAC>=0.16.5
PySocks>=1.7.1
pytest>=8.3.5
pytest-base-url>=2.1.0
pytest-bdd>=8.1.0
pytest-html>=4.1.1
pytest-metadata>=3.1.1
pytest-playwright>=0.7.0
pytest-xdist>=3.6.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
python-slugify>=8.0.4
PyYAML>=6.0.2
requests>=2.32.3
requests-file>=2.1.0
requests-toolbelt>=1.0.0
rich>=13.9.4
rsa>=4.9
s3transfer>=0.11.4
selenium>=4.29.0
six>=1.17.0
smmap>=5.0.2
sniffio>=1.3.1
sortedcontainers>=2.4.0
structlog>=25.2.0
text-unidecode>=1.3
tldextract>=5.1.3
tokenizers>=0.21.1
tomli>=2.2.1
tqdm>=4.67.1
trio>=0.29.0
trio-websocket>=0.12.2
types-requests>=2.31.0.6
types-urllib3>=1.26.25.14
typing-inspect>=0.9.0
typing-inspection>=0.4.0
typing_extensions>=4.12.2
tzdata>=2025.1
urllib3>=1.26.20
wcwidth>=0.2.13
websocket-client>=1.8.0
wrapt>=1.17.2
wsproto>=1.2.0
zipp>=3.21.0
